q1

+++ w1_q1.cpp
@@ -1,10 +1,10 @@
 #include <iostream>
 using namespace std;
 int main() {
-    int n = 0;
+    long n = 0;
     cin >> n;
-    int s = 0;  
-    int t = 0;
+    long s = 0;  
+    long t = 0;
     for (int i = 0; i < (n-1); i++) {
         cin >> t;
         s += t;
